kafka_producer/requirements.txt
praw
click
python-dotenv
confluent_kafka